split scene headers
